moving message read listener to separate hook
